requirements.txt
appdirs==1.4.3
attrs==19.1.0
backcall==0.1.0
cached-property==1.5.1
certifi==2018.11.29
chardet==3.0.4
crispy-forms-materialize==0.2
decorator==4.3.2
defusedxml==0.5.0
Django==2.1.5
django-autoslug==1.9.4
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-js-asset==1.2.2
django-materialize-css==0.0.1
djangorestframework==3.9.2
idna==2.8
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
isodate==0.6.0
jdatetime==3.2.0
jedi==0.13.2
lxml==4.3.3
oauthlib==3.0.1
parso==0.3.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.8
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
sorl-thumbnail==12.5.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
zeep==3.3.1
